rpk: Add 'rpk cluster info', deprecate 'rpk api status'
